refactor: 데이터베이스에 의존성역전 사용
